\name{col.dis}
\docType{data}
\alias{coldis}
\alias{knn10columbus}
\title{Object of class distance for Columbus dataset}

\description{
Distance object, GWT and txt files for columbus dataset
}

\usage{
coldis
}

\format{
A list of neighbors and distances for the columbus neighborhoods}


\keyword{datasets}