\name{utilities}
\alias{fistslsfast}
\alias{fistslserror}
\alias{fifour}
\alias{fierror}
\alias{Ggfastfast}

\alias{arg}
\alias{arg1}
\alias{Omega}
\alias{Omegabis}

\title{Functions used by gstslshet. }

\description{
\itemize{
\item \code{arg} and \code{arg1} are the objective functions of the non-linear 
estimators in the GMM procedure. 
\item \code{Omega} and \code{Omegabis} generates the variance-covariance matrices 
of the Original and Transformed models (See Arraiz et al., 2007 for details.)
\item \code{Ggfastfast} calculates G and g.
\item All other functions perform calculations to estimates various objects 
defined in Appendix B2 and B3 in Arraiz et al., 2007.
}
}






\author{ Gianfranco Piras \email{gpiras@mac.com}}



\keyword{models}
